Refactor everything

+++ workouts/services.py
@@ -1,40 +1,60 @@
 # workouts/services.py
 
 from django.db import transaction
-from .models import Workout, Set
+from django.shortcuts import get_object_or_404
+from rest_framework.exceptions import ValidationError
+from .models import Workout, Exercise, Set
 
 
 @transaction.atomic
-def create_workout(user):
-    new_workout = Workout.objects.create(user=user)
+def create_workout(*, user, week, day):
+    if Workout.objects.filter(user=user, week=week, day=day).exists():
+        raise ValidationError("Workout already exists")
 
-    return new_workout
+    return Workout.objects.create(user=user, week=week, day=day)
 
 
-def create_set(workout, exercise, reps):
-    new_set = Set.objects.create(workout=workout, exercise=exercise, reps=reps)
+@transaction.atomic
+def create_exercise(*, workout, name_of_exercise):
+    new_exercise = Exercise.objects.create(
+        workout=workout, name_of_exercise=name_of_exercise
+    )
+
+    return new_exercise
+
+
+@transaction.atomic
+def create_set(*, exercise, reps):
+    new_set = Set.objects.create(exercise=exercise, reps=reps)
 
     return new_set
 
 
-def create_full_workout(user, sets_data):
-    new_workout = Workout.objects.create(user=user)
+@transaction.atomic
+def create_full_workout(*, user, week, day, exercises_data):
 
-    for set_item in sets_data:
-        Set.objects.create(
-            workout=new_workout, exercise=set_item["exercise"], reps=set_item["reps"]
+    workout = Workout.objects.create(user=user, week=week, day=day)
+
+    for exercise_data in exercises_data:
+        exercise = create_exercise(
+            workout=workout,
+            name_of_exercise=exercise_data["name_of_exercise"]
         )
-    return new_workout
+
+        for set_data in exercise_data["sets"]:
+            create_set(
+                exercise=exercise,
+                reps=set_data["reps"]
+            )
+
+    return workout
 
 
-def update_set(set_instance, **kwargs):
-    for key, value in kwargs.items():
-        setattr(set_instance, key, value)
+def delete_workout(*, workout_id, user):
 
-        set_instance.save()
+    workout = get_object_or_404(Workout, id=workout_id)
 
-        return set_instance
+    if workout.user != user:
+        raise ValidationError("You cannot delete this workout")
 
-
-def delete_workout(workout):
     workout.delete()
